+++ app/(studio)/institution/pathfinder/page.tsx
@@ -1,0 +1,5 @@
+import InstitutionPathfinderClient from "./InstitutionPathfinderClient";
+
+export default function InstitutionPathfinderPage() {
+  return <InstitutionPathfinderClient />;
+}